requirements.txt
discord
aiogoogletrans
python-dotenv
flask
discord-py